Pathways route completed. Fixed lint errors
