+++ lib/models/coursesV2.js
@@ -1,6 +1,4 @@
 const Joi = require('@hapi/joi');
-const glob = require('glob');
-const _ = require('lodash');
 const { Model } = require('objection');
 const ModelBase = require('./helpers/ModelBase');
