+++ lib/routes/pathways.js
@@ -74,6 +74,7 @@ module.exports = [
       handler: async (request, h) => {
         const { pathwayService, pathwayServiceV2, displayService } = request.services();
         const appVersion = request.headers['version-code'] ? request.headers['version-code'] : null;
+
         if (appVersion >= androidVersions.latest) {
           const [err, pathways] = await pathwayServiceV2.find();
           // eslint-disable-next-line
@@ -97,34 +98,33 @@ module.exports = [
 
           logger.info('List of all Pathways');
           return { pathways: pathway };
-        } else {
-          const [err, pathways] = await pathwayService.find();
-          if (err) {
-            logger.error(JSON.stringify(err));
-            return h.response(err).code(err.code);
-          }
-          const pathway = await displayService.pathway(pathways, request.query.courseType);
-
-          _.map(pathway, (p) => {
-            const pathway_lang = [];
-            _.map(p.courses, (course) => {
-              _.map(course.lang_available, (lang) => {
-                if (!_.find(pathway_lang, { code: lang, label: lang_key[lang] })) {
-                  pathway_lang.push({ code: lang, label: lang_key[lang] });
-                }
-              });
-            });
-            p.lang_available = pathway_lang;
-          });
-
-          // if (request.headers.platform !== 'web') {
-          //   logger.info('List of all Pathways');
-          //   return { pathways: _.filter(pathway, (e) => e.code !== 'JVSCPT') };
-          // }
-
-          logger.info('List of all Pathways');
-          return { pathways: pathway };
         }
+        const [err, pathways] = await pathwayService.find();
+        if (err) {
+          logger.error(JSON.stringify(err));
+          return h.response(err).code(err.code);
+        }
+        const pathway = await displayService.pathway(pathways, request.query.courseType);
+
+        _.map(pathway, (p) => {
+          const pathway_lang = [];
+          _.map(p.courses, (course) => {
+            _.map(course.lang_available, (lang) => {
+              if (!_.find(pathway_lang, { code: lang, label: lang_key[lang] })) {
+                pathway_lang.push({ code: lang, label: lang_key[lang] });
+              }
+            });
+          });
+          p.lang_available = pathway_lang;
+        });
+
+        // if (request.headers.platform !== 'web') {
+        //   logger.info('List of all Pathways');
+        //   return { pathways: _.filter(pathway, (e) => e.code !== 'JVSCPT') };
+        // }
+
+        logger.info('List of all Pathways');
+        return { pathways: pathway };
       },
     },
   },
@@ -236,7 +236,7 @@ module.exports = [
       handler: async (request, h) => {
         const { pathwayService, pathwayServiceV2, displayService } = request.services();
         const appVersion = request.headers['version-code'] ? request.headers['version-code'] : null;
-        const courseType = request.query['courseType'] ? request.query['courseType'] : null;
+        const courseType = request.query.courseType ? request.query.courseType : null;
 
         if (appVersion >= androidVersions.latest) {
           const [err, pathway] = await pathwayServiceV2.findById(request.params.pathwayId);
@@ -271,32 +271,67 @@ module.exports = [
         params: Joi.object({
           pathwayId: Pathways.field('id'),
         }),
+        headers: Joi.object({
+          'version-code': Joi.number().integer().optional(),
+        }),
         payload: schemas.pathway,
+        options: { allowUnknown: true },
       },
       handler: async (request, h) => {
-        const { pathwayService, displayService } = request.services();
+        const { pathwayService, pathwayServiceV2, displayService } = request.services();
         const { pathwayId } = request.params;
 
-        const updateAndFetch = async (txn) => {
+        const appVersion = request.headers['version-code'] ? request.headers['version-code'] : null;
+        if (appVersion >= androidVersions.latest) {
           // eslint-disable-next-line
-          const [err, updatedPathway] = await pathwayService.update(
-            pathwayId,
-            request.payload,
-            txn
-          );
+          const [err, pathway] = await pathwayServiceV2.update(pathwayId, request.payload);
           if (err) {
             logger.error(JSON.stringify(err));
-            return [err, null];
+            return h.response(err).code(err.code);
           }
-          return [null, pathwayService.findById(pathwayId, txn)];
-        };
-        const [err, pathway] = await h.context.transaction(updateAndFetch);
+          const [errInFetch, updatedPathway] = await pathwayServiceV2.findById(pathwayId);
+          if (errInFetch) {
+            logger.error(JSON.stringify(errInFetch));
+            return h.response(errInFetch).code(errInFetch.code);
+          }
+          logger.info(`Pathway id- ${pathwayId} updated`);
+          return { pathway: await displayService.pathwayV2(updatedPathway) };
+        }
+
+        // eslint-disable-next-line
+        const [err, pathway] = await pathwayService.update(pathwayId, request.payload);
         if (err) {
           logger.error(JSON.stringify(err));
           return h.response(err).code(err.code);
         }
+        const [errInFetch, updatedPathway] = await pathwayService.findById(pathwayId);
+        if (errInFetch) {
+          logger.error(JSON.stringify(errInFetch));
+          return h.response(errInFetch).code(errInFetch.code);
+        }
         logger.info(`Pathway id- ${pathwayId} updated`);
-        return { pathway: await displayService.pathway(pathway) };
+        return { pathway: await displayService.pathway(updatedPathway) };
+
+        // const updateAndFetch = async (txn) => {
+        //   // eslint-disable-next-line
+        //   const [err, updatedPathway] = await pathwayService.update(
+        //     pathwayId,
+        //     request.payload,
+        //     txn
+        //   );
+        //   if (err) {
+        //     logger.error(JSON.stringify(err));
+        //     return [err, null];
+        //   }
+        //   return [null, pathwayService.findById(pathwayId, txn)];
+        // };
+        // const [err, pathway] = await h.context.transaction(updateAndFetch);
+        // if (err) {
+        //   logger.error(JSON.stringify(err));
+        //   return h.response(err).code(err.code);
+        // }
+        // logger.info(`Pathway id- ${pathwayId} updated`);
+        // return { pathway: await displayService.pathway(pathway) };
       },
     },
   },
@@ -314,9 +349,23 @@ module.exports = [
         params: Joi.object({
           pathwayId: Pathways.field('id'),
         }),
+        headers: Joi.object({
+          'version-code': Joi.number().integer().optional(),
+        }),
+        options: { allowUnknown: true },
       },
       handler: async (request, h) => {
-        const { pathwayService } = request.services();
+        const { pathwayService, pathwayServiceV2 } = request.services();
+        const appVersion = request.headers['version-code'] ? request.headers['version-code'] : null;
+        if (appVersion >= androidVersions.latest) {
+          const [err, deleted] = await pathwayServiceV2.deleteById(request.params.pathwayId);
+          if (err) {
+            logger.error(JSON.stringify(err));
+            return h.response(err).code(err.code);
+          }
+          logger.info(`Delete a pathway by pathway id- ${request.params.pathwayId}`);
+          return deleted;
+        }
         const [err, deleted] = await pathwayService.deleteById(request.params.pathwayId);
         if (err) {
           logger.error(JSON.stringify(err));
@@ -486,11 +535,37 @@ module.exports = [
         query: Joi.object({
           courseType: Joi.string().valid('markdown', 'json').allow(null).optional(),
         }),
+        headers: Joi.object({
+          'version-code': Joi.number().integer().optional(),
+        }),
+        options: { allowUnknown: true },
       },
       handler: async (request, h) => {
-        const { pathwayService, displayService } = request.services();
+        const { pathwayService, pathwayServiceV2, displayService } = request.services();
+        const appVersion = request.headers['version-code'] ? request.headers['version-code'] : null;
+        if (appVersion >= androidVersions.latest) {
+          const [err, pathway] = await pathwayServiceV2.getDefaultPathway();
+          if (err) {
+            logger.error(JSON.stringify(err));
+            return h.response(err).code(err.code);
+          }
+          const [errInPathway, pathwayCourses] = await displayService.getPathwayCoursesV2(
+            pathway.id
+          );
+          if (errInPathway) return h.response(errInPathway).code(errInPathway.code);
+          const pathway_lang = [];
+          _.map(pathwayCourses.courses, (p) => {
+            _.map(p.lang_available, (lang) => {
+              if (!_.find(pathway_lang, { code: lang, label: lang_key[lang] })) {
+                pathway_lang.push({ code: lang, label: lang_key[lang] });
+              }
+            });
+          });
+          pathwayCourses.lang_available = pathway_lang;
+          logger.info('Get all courses of default pathway');
+          return pathwayCourses;
+        }
         const [err, pathway] = await pathwayService.getDefaultPathway();
-
         if (err) {
           logger.error(JSON.stringify(err));
           return h.response(err).code(err.code);
@@ -544,10 +619,34 @@ module.exports = [
         query: Joi.object({
           courseType: Joi.string().valid('markdown', 'json').allow(null).optional(),
         }),
+        headers: Joi.object({
+          'version-code': Joi.number().integer().optional(),
+        }),
+        options: { allowUnknown: true },
       },
       handler: async (request, h) => {
         const { displayService } = request.services();
         const { pathwayId } = request.params;
+        const appVersion = request.headers['version-code'] ? request.headers['version-code'] : null;
+
+        if (appVersion >= androidVersions.latest) {
+          const [errInPathway, pathwayCourses] = await displayService.getPathwayCoursesV2(
+            pathwayId
+          );
+          if (errInPathway) return h.response(errInPathway).code(errInPathway.code);
+          const pathway_lang = [];
+          _.map(pathwayCourses.courses, (p) => {
+            _.map(p.lang_available, (lang) => {
+              if (!_.find(pathway_lang, { code: lang, label: lang_key[lang] })) {
+                pathway_lang.push({ code: lang, label: lang_key[lang] });
+              }
+            });
+          });
+          pathwayCourses.lang_available = pathway_lang;
+          logger.info('Get all courses of a particular pathway id');
+          return pathwayCourses;
+        }
+
         const [errInPathway, pathwayCourses] = await displayService.getPathwayCourses(
           pathwayId,
           request.query.courseType
@@ -597,11 +696,47 @@ module.exports = [
         payload: Joi.object({
           courseIds: Joi.array().items(PathwayCourses.field('course_id')),
         }),
+        headers: Joi.object({
+          'version-code': Joi.number().integer().optional(),
+        }),
+        options: { allowUnknown: true },
       },
       handler: async (request, h) => {
-        const { pathwayService, displayService } = request.services();
+        const { pathwayService, pathwayServiceV2, displayService } = request.services();
         const { pathwayId } = request.params;
         const { courseIds } = request.payload;
+
+        const appVersion = request.headers['version-code'] ? request.headers['version-code'] : null;
+        if (appVersion >= androidVersions.latest) {
+          const updateAndFetch = async () => {
+            const [err, updatedPathway] = await pathwayServiceV2.upsertPathwayCourses(
+              pathwayId,
+              courseIds
+            );
+            if (err) {
+              logger.error(JSON.stringify(err));
+              return h.response(err).code(err.code);
+            }
+            return updatedPathway;
+          };
+
+          const checkIfUpdated = await h.context.transaction(updateAndFetch);
+
+          if ('error' in checkIfUpdated) {
+            logger.error(JSON.stringify(checkIfUpdated));
+            return checkIfUpdated;
+          }
+          const [errInPathway, pathwayCourses] = await displayService.getPathwayCoursesV2(
+            pathwayId
+          );
+          if (errInPathway) {
+            return h.response(errInPathway).code(errInPathway.code);
+          }
+          logger.info(
+            `id- ${request.auth.credentials.id} Add, update and delete courses in a pathway`
+          );
+          return pathwayCourses;
+        }
 
         const updateAndFetch = async () => {
           const [err, updatedPathway] = await pathwayService.upsertPathwayCourses(
@@ -643,9 +778,27 @@ module.exports = [
         params: Joi.object({
           pathwayId: PathwayCompletion.field('pathway_id'),
         }),
+        headers: Joi.object({
+          'version-code': Joi.number().integer().optional(),
+        }),
+        options: { allowUnknown: true },
       },
       handler: async (request, h) => {
-        const { pathwayService } = request.services();
+        const { pathwayService, pathwayServiceV2 } = request.services();
+        const appVersion = request.headers['version-code'] ? request.headers['version-code'] : null;
+
+        if (appVersion >= androidVersions.latest) {
+          const [err, marked] = await pathwayServiceV2.markPathwayComplete(
+            request.auth.credentials.id,
+            request.params.pathwayId
+          );
+          if (err) {
+            logger.error(JSON.stringify(err));
+            return h.response(err).code(err.code);
+          }
+          logger.info(`id- ${request.auth.credentials.id} Mark pathway completion`);
+          return marked;
+        }
         const [err, marked] = await pathwayService.markPathwayComplete(
           request.auth.credentials.id,
           request.params.pathwayId
@@ -670,9 +823,27 @@ module.exports = [
         params: Joi.object({
           pathwayId: PathwayCompletion.field('pathway_id'),
         }),
+        headers: Joi.object({
+          'version-code': Joi.number().integer().optional(),
+        }),
+        options: { allowUnknown: true },
       },
       handler: async (request, h) => {
-        const { pathwayService } = request.services();
+        const { pathwayService, pathwayServiceV2 } = request.services();
+
+        const appVersion = request.headers['version-code'] ? request.headers['version-code'] : null;
+        if (appVersion >= androidVersions.latest) {
+          const [err, unmarked] = await pathwayServiceV2.removePathwayComplete(
+            request.auth.credentials.id,
+            request.params.pathwayId
+          );
+          if (err) {
+            logger.error(JSON.stringify(err));
+            return h.response(err).code(err.code);
+          }
+          logger.info(`id- ${request.auth.credentials.id} Unmark pathway completion`);
+          return unmarked;
+        }
         const [err, unmarked] = await pathwayService.removePathwayComplete(
           request.auth.credentials.id,
           request.params.pathwayId
@@ -693,8 +864,26 @@ module.exports = [
       description: 'Get all completed pathways',
       tags: ['api'],
       auth: { strategy: 'jwt' },
+      validate: {
+        headers: Joi.object({
+          'version-code': Joi.number().integer().optional(),
+        }),
+        options: { allowUnknown: true },
+      },
       handler: async (request, h) => {
-        const { pathwayService } = request.services();
+        const { pathwayService, pathwayServiceV2 } = request.services();
+        const appVersion = request.headers['version-code'] ? request.headers['version-code'] : null;
+        if (appVersion >= androidVersions.latest) {
+          const [err, completed] = await pathwayServiceV2.getPathwayComplete(
+            request.auth.credentials.id
+          );
+          if (err) {
+            logger.error(JSON.stringify(err));
+            return h.response(err).code(err.code);
+          }
+          logger.info(`id- ${request.auth.credentials.id} Get all completed pathways`);
+          return completed;
+        }
         const [err, completed] = await pathwayService.getPathwayComplete(
           request.auth.credentials.id
         );
