+++ lib/services/display.js
@@ -25,7 +25,9 @@ module.exports = class DisplayService extends Schmervice.Service {
     });
     return transformedData;
   };
+
   _applyTransform = (results, func) => (_.isArray(results) ? _.map(results, func) : func(results));
+
   userProfile = async (user) => {
     const { User } = this.server.models();
     const results = await User.fetchGraph(user, { roles: true, pathways: true });
@@ -34,6 +36,7 @@ module.exports = class DisplayService extends Schmervice.Service {
 
   _applyTransformV2 = (results, func) =>
     _.isArray(results) ? _.map(results, func) : func(results);
+
   userProfile = async (user) => {
     const { User } = this.server.models();
     const results = await User.fetchGraph(user, { roles: true, pathwaysV2: true });
@@ -77,13 +80,10 @@ module.exports = class DisplayService extends Schmervice.Service {
       courses: true,
     });
 
-    console.log(course_type, 'KKKKKKOoooo');
-
     pathways = pathways.length ? pathways : [pathways];
     pathways = _.map(pathways, (p) => {
       p.courses = _.filter(p.courses, (c) => {
         if (course_type === 'json') {
-          console.log('will be');
           return c.course_type === 'json';
         }
         return c.course_type !== 'json';
