+++ lib/services/pathway.js
@@ -71,7 +71,7 @@ module.exports = class PathwayService extends Schmervice.Service {
       info.tracking_days_lock_before_cycle = null;
     }
     try {
-      await Pathways.query(txn).update(info).where({ id });
+      await Pathways.query(txn).patch(info).where({ id });
       return [null, id];
     } catch (err) {
       return [errorHandler(err), null];
